refactor: :art: Move jwt-models to new internal package jwtmodels as not required to be exported in this plugin package
* Claims IsValidForRole now exported and we pass a debug logger in
